requirements.txt
streamlit
langchain
langchain-community
ollama
pandas
fpdf2
python-dateutil
ipwhois
pytest
